% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LidaROllama.R
\docType{class}
\name{LidaROllama}
\alias{LidaROllama}
\title{LidaROllama: An R6 Class for Interacting with the Ollama API}
\value{
An instance of the `LidaROllama` class.
}
\description{
`LidaROllama` is an R6 class for generating text using the Ollama API.
}
\examples{
\dontrun{
api <- LidaROllama$new(
  model_name = "llama3:latest",
  temperature = 0.7,
  max_length = 512,
  sysprompt = "Feluda",
  api_url = "http://localhost:11434/api/generate"
)
response <- api$ollama_api("What is R?")
print(response)
}
}
\keyword{package}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LidaROllama-new}{\code{LidaROllama$new()}}
\item \href{#method-LidaROllama-ollama_api}{\code{LidaROllama$ollama_api()}}
\item \href{#method-LidaROllama-clone}{\code{LidaROllama$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LidaROllama-new"></a>}}
\if{latex}{\out{\hypertarget{method-LidaROllama-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LidaROllama$new(model_name, temperature, max_length, sysprompt, api_url)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LidaROllama-ollama_api"></a>}}
\if{latex}{\out{\hypertarget{method-LidaROllama-ollama_api}{}}}
\subsection{Method \code{ollama_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LidaROllama$ollama_api(prompt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LidaROllama-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LidaROllama-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LidaROllama$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
